pyproject.toml
[project]
name = "strava-viewer"
version = "0.4.0"
description = ""
authors = [{ name = "Ariel Parra", email = "darkaico@gmail.com" }]
requires-python = ">=3.11,<4"
dependencies = [
    "requests>=2.28",
    "dacite>=1.6",
    "marshmallow>=3.14",
    "python-dotenv>=0.21",
    "flask>=2.2",
    "Flask-WTF>=1.0",
    "redis>=4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2",
    "pytest-mock>=3.10",
    "faker>=15.3",
    "pytest-cov>=4.0",
    "ipdb>=0.13",
    "black>=22.10",
    "pre-commit>=2.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["strava_viewer"]

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
